+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # adrez-mapper Revisions
 
+2020-06-02
+- add: auto array for record elements
+
 2020-06-29
 - add: name parser can be configured by a config. Keys: suffix, prefix, titles 
 - fix: campaign.group (groupGuid, group, groupId). Callback: campaignGroup